Partie 1 : Classes data et services séparées
