+++ src/core/data/Account.java
@@ -1,4 +1,4 @@
-package core;
+package core.data;
 
 import estorePojo.exceptions.InsufficientBalanceException;
