+++ src/core/data/Order.java
@@ -1,12 +1,12 @@
-package core;
+package core.data;
 
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
+import core.services.Client;
 import estorePojo.exceptions.UnknownItemException;
 
 public class Order {
